requirements.txt
pdfplumber>=0.11.0
pandas>=2.1.0
openpyxl>=3.1.0
openai>=1.10.0
python-dotenv>=1.0.0
